+++ pkg/scaling/scalers_builder.go
@@ -252,6 +252,8 @@ func buildScaler(ctx context.Context, client client.Client, triggerType string, 
 		return scalers.NewSeleniumGridScaler(config)
 	case "solace-event-queue":
 		return scalers.NewSolaceScaler(config)
+	case "solarwinds":
+		return scalers.NewSolarWindsScaler(config)
 	case "solr":
 		return scalers.NewSolrScaler(config)
 	case "splunk":